Fix bug 1098194 - Move config_lazy helper to demo models module to prevent circular import.
This happens since the utility requires importing the FIELD list from the constance app, which triggers the population of the Django app cache, which imports other models and voila you have a circular implicite import error.

This would not happen when using modern Django.
